+++ weni/grpc/org/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "weni.grpc.org.apps.OrgGrpcConfig"
