+++ weni/grpc/org/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class OrgGrpcConfig(AppConfig):
-    name = "weni.org_grpc"
+    name = "weni.grpc.org"
