+++ weni/grpc/org/tests.py
@@ -5,8 +5,8 @@ from django_grpc_framework.test import FakeRpcError, RPCTransactionTestCase
 from rest_framework.exceptions import ValidationError
 
 from temba.orgs.models import Org
-from weni.org_grpc.grpc_gen import org_pb2, org_pb2_grpc
-from weni.org_grpc.serializers import SerializerUtils
+from weni.grpc.org.grpc_gen import org_pb2, org_pb2_grpc
+from weni.grpc.org.serializers import SerializerUtils
 
 
 class OrgServiceTest(RPCTransactionTestCase):
